# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-flow",
    "version": "1",
    "description": "Workflow orchestration. Definitions are declarative JSON; workers are language-agnostic and lease tasks from a queue."
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived, revocable, scoped. For CLIs and CI."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived token from POST /v1/auth/token. For worker fleets: a captured token is useless within the hour."
      }
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/hooks/{id}": {
      "post": {
        "operationId": "hookReceiver_receive",
        "summary": "Receive a webhook delivery",
        "tags": [
          "webhooks"
        ],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/v1/ns/{ns}/incoming-webhooks": {
      "get": {
        "operationId": "incomingWebhook_list",
        "summary": "List inbound webhooks",
        "tags": [
          "webhooks"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "post": {
        "operationId": "incomingWebhook_create",
        "summary": "Create an inbound webhook",
        "tags": [
          "webhooks"
        ],
        "x-required-scopes": [
          "workflows:write"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "verifier": {
                    "type": "string",
                    "enum": [
                      "GITHUB",
                      "STRIPE",
                      "SLACK",
                      "SHOPIFY",
                      "HMAC",
                      "HEADER",
                      "NONE"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "header": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "algorithm": {
                        "type": "string",
                        "enum": [
                          "sha256",
                          "sha1",
                          "sha512"
                        ]
                      },
                      "encoding": {
                        "type": "string",
                        "enum": [
                          "hex",
                          "base64"
                        ]
                      },
                      "prefix": {
                        "type": "string",
                        "maxLength": 50
                      }
                    }
                  },
                  "secretName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "verifier"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/incoming-webhooks/{name}": {
      "delete": {
        "operationId": "incomingWebhook_remove",
        "summary": "Delete an inbound webhook",
        "tags": [
          "webhooks"
        ],
        "x-required-scopes": [
          "workflows:write"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "get": {
        "operationId": "incomingWebhook_get",
        "summary": "Fetch an inbound webhook",
        "tags": [
          "webhooks"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "put": {
        "operationId": "incomingWebhook_update",
        "summary": "Update an inbound webhook",
        "tags": [
          "webhooks"
        ],
        "x-required-scopes": [
          "workflows:write"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "verifier": {
                    "type": "string",
                    "enum": [
                      "GITHUB",
                      "STRIPE",
                      "SLACK",
                      "SHOPIFY",
                      "HMAC",
                      "HEADER",
                      "NONE"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "header": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                      },
                      "algorithm": {
                        "type": "string",
                        "enum": [
                          "sha256",
                          "sha1",
                          "sha512"
                        ]
                      },
                      "encoding": {
                        "type": "string",
                        "enum": [
                          "hex",
                          "base64"
                        ]
                      },
                      "prefix": {
                        "type": "string",
                        "maxLength": 50
                      }
                    }
                  },
                  "secretName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "verifier"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
  }
}